+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2021-07-29
+- Revamp of FormH procedure to use split comparison and calculation stages to save memory
+- Many small bug fixes regarding memory
+- Progress bar fixed
+
 ## [0.3.6] - 2021-07-14
 - Removed usage of MPI windows - suspected memory leak due to pointers not being deallocated
 - FormH bug fix - initial workload distribution was missing a chunk of determinants